refactor: separate parser logic and cli args more clearly
