+++ lib/cli-args.js
@@ -1,54 +1,26 @@
-import fs from 'fs';
-import _ from 'lodash';
 import { DEFAULT_BASE_PATH } from 'appium-base-driver';
+import { parseSecurityFeatures, parseDefaultCaps, parseInstallTypes } from './parser-helpers';
+import { INSTALL_TYPES, DEFAULT_APPIUM_HOME } from './driver-config';
 
-// serverArgs will be added to the `server` (default) subcommand
-function parseSecurityFeatures (features) {
-  const splitter = (splitOn, str) => `${str}`.split(splitOn).map((s) => s.trim()).filter(Boolean);
-  let parsedFeatures;
-  try {
-    parsedFeatures = splitter(',', features);
-  } catch (err) {
-    throw new Error('Could not parse value of --allow/deny-insecure. Should be ' +
-                    'a list of strings separated by commas, or a path to a file ' +
-                    'listing one feature name per line.');
-  }
+// sharedArgs will be added to every subcommand
+const sharedArgs = [
+  [['-ah', '--home', '--appium-home'], {
+    required: false,
+    defaultValue: process.env.APPIUM_HOME || DEFAULT_APPIUM_HOME,
+    help: 'The path to the directory where Appium will keep installed drivers, plugins, and any other metadata necessary for its operation',
+    dest: 'appiumHome',
+  }],
 
-  if (parsedFeatures.length === 1 && fs.existsSync(parsedFeatures[0])) {
-    // we might have a file which is a list of features
-    try {
-      const fileFeatures = fs.readFileSync(parsedFeatures[0], 'utf8');
-      parsedFeatures = splitter('\n', fileFeatures);
-    } catch (err) {
-      throw new Error(`Attempted to read --allow/deny-insecure feature names ` +
-                      `from file ${parsedFeatures[0]} but got error: ${err.message}`);
-    }
-  }
+  [['--log-filters'], {
+    dest: 'logFilters',
+    defaultValue: null,
+    required: false,
+    help: 'Set the full path to a JSON file containing one or more log filtering rules',
+    example: '/home/rules.json',
+  }],
+];
 
-  return parsedFeatures;
-}
-
-function parseDefaultCaps (caps) {
-  try {
-    // use synchronous file access, as `argparse` provides no way of either
-    // awaiting or using callbacks. This step happens in startup, in what is
-    // effectively command-line code, so nothing is blocked in terms of
-    // sessions, so holding up the event loop does not incur the usual
-    // drawbacks.
-    if (fs.statSync(caps).isFile()) {
-      caps = fs.readFileSync(caps, 'utf8');
-    }
-  } catch (err) {
-    // not a file, or not readable
-  }
-  caps = JSON.parse(caps);
-  if (!_.isPlainObject(caps)) {
-    throw 'Invalid format for default capabilities';
-  }
-  return caps;
-}
-
-export default [
+const serverArgs = [
   [['--shell'], {
     required: false,
     defaultValue: null,
@@ -346,3 +318,77 @@ export default [
   }],
 ];
 
+const globalDriverArgs = [
+  [['--json'], {
+    required: false,
+    defaultValue: false,
+    action: 'storeTrue',
+    help: 'Use JSON for output format',
+    nargs: 0,
+    dest: 'json'
+  }]
+];
+
+const driverListArgs = [
+  ...globalDriverArgs,
+  [['--installed'], {
+    required: false,
+    defaultValue: false,
+    action: 'storeTrue',
+    help: 'List only installed drivers',
+    nargs: 0,
+    dest: 'showInstalled'
+  }],
+  [['--updates'], {
+    required: false,
+    defaultValue: false,
+    action: 'storeTrue',
+    help: 'Show information about newer versions',
+    nargs: 0,
+    dest: 'showUpdates'
+  }]
+];
+
+const driverInstallArgs = [
+  ...globalDriverArgs,
+  [['driver'], {
+    type: 'string',
+    example: 'xcuitest',
+    help: 'Name of the driver to install',
+  }],
+  [['--source'], {
+    required: false,
+    defaultValue: null,
+    type: parseInstallTypes,
+    help: `Where to look for the driver if it is not one of Appium's verified ` +
+          `drivers. Possible values: ${JSON.stringify(INSTALL_TYPES)}`,
+    dest: 'installType'
+  }],
+];
+
+const driverUninstallArgs = [
+  ...globalDriverArgs,
+  [['driver'], {
+    type: 'string',
+    example: 'xcuitest',
+    help: 'Name of the driver to uninstall',
+  }],
+];
+
+const driverUpdateArgs = [
+  ...globalDriverArgs,
+  [['driver'], {
+    type: 'string',
+    example: 'xcuitest',
+    help: 'Name of the driver to update, or the word "all" to update all. To see available updates, run "appium driver list --installed --updates"',
+  }],
+];
+
+export {
+  sharedArgs,
+  serverArgs,
+  driverListArgs,
+  driverInstallArgs,
+  driverUninstallArgs,
+  driverUpdateArgs,
+};
