+++ lib/driver-config.js
@@ -4,6 +4,8 @@ import { fs, mkdirp } from 'appium-support';
 import path from 'path';
 import YAML from 'yaml';
 
+const DEFAULT_APPIUM_HOME = path.resolve(process.env.HOME, '.appium');
+
 const CONFIG_FILE_NAME = 'drivers.yaml';
 const CONFIG_SCHEMA_REV = 1;
 
@@ -195,5 +197,5 @@ export default class DriverConfig {
 
 export {
   INSTALL_TYPE_NPM, INSTALL_TYPE_GIT, INSTALL_TYPE_LOCAL, INSTALL_TYPE_GITHUB,
-  INSTALL_TYPES
+  INSTALL_TYPES, DEFAULT_APPIUM_HOME
 };
