+++ lib/main.js
@@ -6,8 +6,7 @@ import logger from './logger'; // logger needs to remain first of imports
 import _ from 'lodash';
 import { server as baseServer, routeConfiguringFunction as makeRouter } from 'appium-base-driver';
 import { asyncify } from 'asyncbox';
-import { default as getParser } from './parser';
-import { getDefaultServerArgs } from './server-parser';
+import { default as getParser, getDefaultServerArgs } from './parser';
 import { logger as logFactory, util } from 'appium-support';
 import {
   showConfig, checkNodeOk, validateServerArgs,
