+++ lib/parser.js
@@ -1,29 +1,9 @@
 import path from 'path';
 import _ from 'lodash';
 import { ArgumentParser } from 'argparse';
-import { addDriverToParser } from './driver-parser';
-import { addServerToParser } from './server-parser';
+import { sharedArgs, serverArgs, driverListArgs, driverInstallArgs, driverUninstallArgs,
+         driverUpdateArgs } from './cli-args';
 import { rootDir } from './utils';
-
-const DEFAULT_APPIUM_HOME = path.resolve(process.env.HOME, '.appium');
-
-// sharedArgs will be added to every subcommand
-const sharedArgs = [
-  [['-ah', '--home', '--appium-home'], {
-    required: false,
-    defaultValue: process.env.APPIUM_HOME || DEFAULT_APPIUM_HOME,
-    help: 'The path to the directory where Appium will keep installed drivers, plugins, and any other metadata necessary for its operation',
-    dest: 'appiumHome',
-  }],
-
-  [['--log-filters'], {
-    dest: 'logFilters',
-    defaultValue: null,
-    required: false,
-    help: 'Set the full path to a JSON file containing one or more log filtering rules',
-    example: '/home/rules.json',
-  }],
-];
 
 function getParser (debug = false) {
   const parser = new ArgumentParser({
@@ -59,5 +39,58 @@ function getParser (debug = false) {
   return parser;
 }
 
+function addServerToParser (sharedArgs, subParsers, debug) {
+  const serverParser = subParsers.addParser('server', {
+    addHelp: true,
+    help: 'Run an Appium server',
+    debug
+  });
+
+  for (const [flags, opts] of [...sharedArgs, ...serverArgs]) {
+    // addArgument mutates arguments so make copies
+    serverParser.addArgument([...flags], {...opts});
+  }
+
+  return serverArgs;
+}
+
+function getDefaultServerArgs () {
+  let defaults = {};
+  for (let [, arg] of serverArgs) {
+    defaults[arg.dest] = arg.defaultValue;
+  }
+  return defaults;
+}
+
+function addDriverToParser (sharedArgs, subParsers, debug) {
+  const driverParser = subParsers.addParser('driver', {
+    addHelp: true,
+    help: 'Access the driver management CLI commands',
+    debug
+  });
+  const driverSubParsers = driverParser.addSubparsers({
+    dest: 'driverCommand',
+    debug
+  });
+  const parserSpecs = [
+    {command: 'list', args: driverListArgs,
+     help: 'List available and installed drivers'},
+    {command: 'install', args: driverInstallArgs,
+     help: 'Install a driver'},
+    {command: 'uninstall', args: driverUninstallArgs,
+     help: 'Uninstall a driver'},
+    {command: 'update', args: driverUpdateArgs,
+     help: 'Update installed drivers to the latest version'},
+  ];
+
+  for (const {command, args, help} of parserSpecs) {
+    const parser = driverSubParsers.addParser(command, {help, debug});
+    for (const [flags, opts] of [...sharedArgs, ...args]) {
+      // addArgument mutates params so make sure to send in copies instead
+      parser.addArgument([...flags], {...opts});
+    }
+  }
+}
+
 export default getParser;
-export { getParser, DEFAULT_APPIUM_HOME };
+export { getParser, getDefaultServerArgs };
